에러코드 체계 적용

+++ src/modules/classification/classification.service.ts
@@ -1,26 +1,15 @@
-import { Injectable } from '@nestjs/common';
-import {
-  Folder,
-  Post,
-  AIClassification,
-} from '@src/infrastructure/database/schema';
+import { BadRequestException, Injectable } from '@nestjs/common';
 
-import { InjectModel } from '@nestjs/mongoose';
-import { Model, Types } from 'mongoose';
-import { ClassficiationRepository } from './classification.repository';
-import { PostsRepository } from '../posts/posts.repository';
-import {
-  ClassificationFolderWithCount,
-  PostListInClassificationFolder,
-} from './dto/classification.dto';
 import { PaginationQuery, sum } from '@src/common';
-import { PostListInFolderResponse } from '../folders/responses';
+import { Types } from 'mongoose';
+import { PostsRepository } from '../posts/posts.repository';
+import { ClassficiationRepository } from './classification.repository';
+import { ClassificationFolderWithCount } from './dto/classification.dto';
+import { C001 } from './error';
 
 @Injectable()
 export class ClassificationService {
   constructor(
-    @InjectModel(Folder.name) private folderModel: Model<Folder>,
-    @InjectModel(Post.name) private postModel: Model<Post>,
     private readonly classficationRepository: ClassficiationRepository,
     private readonly postRepository: PostsRepository,
   ) {}
@@ -105,9 +94,22 @@ export class ClassificationService {
   }
 
   async abortClassification(userId: string, postId: string) {
-    const post = await this.postModel.findById(postId).exec();
+    const post = await this.postRepository.findPostOrThrow({
+      _id: postId,
+    });
 
-    await this.classficationRepository.delete(
+    if (!post.aiClassificationId) {
+      throw new BadRequestException(C001);
+    }
+
+    const classification = await this.classficationRepository.findById(
+      post.aiClassificationId.toString(),
+    );
+    if (!classification) {
+      throw new BadRequestException(C001);
+    }
+
+    return await this.classficationRepository.delete(
       post.aiClassificationId.toString(),
     );
   }
